% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelc.r
\name{build_additive_term}
\alias{build_additive_term}
\title{Get SQL representing a continuous term in the model with no interactions}
\usage{
build_additive_term(model, additive_term, first = FALSE)
}
\arguments{
\item{model}{A list with the same signature as the output of \code{lm} or \code{glm}}

\item{additive_term}{A parameter name.}

\item{first}{A logical flag signaling whether the term is the first term in the formula}
}
\value{
A SQL character string representing an additive term
}
\description{
Get SQL representing a continuous term in the model with no interactions
}
