% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelc.r
\name{build_factor_case_statements}
\alias{build_factor_case_statements}
\title{Build SQL CASE statements representing the factors in the model}
\usage{
build_factor_case_statements(model, first = FALSE)
}
\arguments{
\item{model}{A list with the same signature as the output of \code{lm} or \code{glm}}

\item{first}{A logical flag signaling whether the term is the first term in the formula}
}
\value{
A character string representing a SQL CASE statement
}
\description{
Build SQL CASE statements representing the factors in the model
}
