% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelc.r
\name{build_interaction_term}
\alias{build_interaction_term}
\title{Build a SQL interaction term}
\usage{
build_interaction_term(model, interaction_term, first = FALSE)
}
\arguments{
\item{model}{A list with the same signature as the output of \code{lm} or \code{glm}}

\item{interaction_term}{The raw interaction term (a character string) from the R model}

\item{first}{A logical flag signaling whether the term is the first term in the formula}
}
\value{
A character string representing a SQL interaction term
}
\description{
Build a SQL interaction term
}
