% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelc.r
\name{build_intercept}
\alias{build_intercept}
\title{Get SQL representing the intercept term given the R model and parameter name}
\usage{
build_intercept(model, parameter, first = FALSE)
}
\arguments{
\item{model}{A list with the same signature as the output of \code{lm} or \code{glm}}

\item{parameter}{A parameter name.}

\item{first}{A logical flag signaling whether the term is the first term in the formula}
}
\value{
A SQL character string representing the intercept term in the model
}
\description{
Get SQL representing the intercept term given the R model and parameter name
}
