% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelc.r
\name{build_product}
\alias{build_product}
\title{Build a SQL product}
\usage{
build_product(lhs, rhs)
}
\arguments{
\item{lhs}{A character string representing the left hand side of the multiplication}

\item{rhs}{A character string representing the right hand side of the multiplication}
}
\value{
A character string representing a valid SQL product term
}
\description{
Build a SQL product
}
