% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelc.r
\name{extract_level}
\alias{extract_level}
\title{Extract the level from the factor name}
\usage{
extract_level(parameter, factor)
}
\arguments{
\item{parameter}{A parameter name}

\item{factor}{A factor term}
}
\value{
A SQL string literal representing the factor level
}
\description{
Extract the level from the factor name
}
