% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelc.r
\name{extract_parameter_coefficient}
\alias{extract_parameter_coefficient}
\title{Extract the coefficient of a model parameter}
\usage{
extract_parameter_coefficient(model, parameter)
}
\arguments{
\item{model}{A list with the same signature as the output of \code{lm} or \code{glm}}

\item{parameter}{A character string corresponding to a model predictor}
}
\value{
A double corresponding to the coefficient, or 0 if the coefficient is missing
}
\description{
Extract the coefficient of a model parameter
}
