% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelc.r
\name{extract_parameters}
\alias{extract_parameters}
\title{Extract parameters from a linear model}
\usage{
extract_parameters(model)
}
\arguments{
\item{model}{A list with the same signature as the output of \code{lm} or \code{glm}}
}
\value{
A character vector of terms from a linear model
}
\description{
Extract parameters from a linear model
}
