% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelc.r
\name{has_parameter}
\alias{has_parameter}
\title{Check if an R model contains a coefficient}
\usage{
has_parameter(model, parameter)
}
\arguments{
\item{model}{A list with the same signature as the output of \code{lm} or \code{glm}}

\item{parameter}{A parameter name}
}
\value{
A logical representing whether a coefficient is present in the model
}
\description{
Check if an R model contains a coefficient
}
