% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelc.r
\name{is_factor}
\alias{is_factor}
\title{Detect if the given model term is a factor}
\usage{
is_factor(parameter, model)
}
\arguments{
\item{parameter}{A parameter name.}

\item{model}{A list with the same signature as the output of \code{lm} or \code{glm}}
}
\value{
A logical representing whether or not the term is a factor
}
\description{
Detect if the given model term is a factor
}
