% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelc.r
\name{is_interaction}
\alias{is_interaction}
\title{Detect if the given model term is an interaction}
\usage{
is_interaction(parameter)
}
\arguments{
\item{parameter}{A parameter name.}
}
\value{
A logical representing whether or not the term is an interaction
}
\description{
Detect if the given model term is an interaction
}
