% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelc.r
\name{is_intercept}
\alias{is_intercept}
\title{Check if the given parameter is the intercept}
\usage{
is_intercept(parameter)
}
\arguments{
\item{parameter}{A parameter name.}
}
\value{
A logical representing whether the given parameter is the intercept
}
\description{
Check if the given parameter is the intercept
}
