% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resample_bootstrap.R
\name{Resample_bootstrap}
\alias{Resample_bootstrap}
\title{bootstrap a resample object}
\usage{
Resample_bootstrap(data)
}
\arguments{
\item{data}{The \code{data} should be a data.frame}
}
\value{

}
\description{
\code{Resample_bootstrap()} bootstrap the subset for a data.frame, using resample modal
class. Rows are sampled for data.
}
\examples{
b <- resample_bootstrap(mtcars)
b
as.integer(b)
as.data.frame(b)
lm(mpg ~ wt, data = b)
}
\seealso{
Other resampling techniques:\code{\link[=resample(), resample_partition]{resample(), resample_partition()}}
}
