% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resample_partition.R
\name{Resample_partition}
\alias{Resample_partition}
\title{partition a resample object}
\usage{
Resample_partition(data, p)
}
\arguments{
\item{data}{The \code{data} should be a data.frame}

\item{p}{The \code{p} represent partition}
}
\value{
resample
}
\description{
\code{Resample_partition()} A data.frame were splited test and train subset data
}
\examples{
ex <- Resample_partition(mtcars, c(test = 0.3, train = 0.7))
}
\seealso{
Other resampling techniques:\code{\link[=Resample(), Resample_partition]{Resample(), Resample_partition()}}
}
