% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resample.R
\name{Resample}
\alias{Resample}
\title{Create a resample object}
\usage{
Resample(data, idx)
}
\arguments{
\item{data}{The \code{data} should be a data.frame}

\item{idx}{Index to choose need rows of data}
}
\value{
vlaue as resample objects
}
\description{
\code{Resample()} creates the subset for a data.frame, using resample modal
class. Rows are typically choosed for data.
}
\examples{
resample(mtcars, 1:10)
}
\references{
modelr
}
\seealso{
Other resampling techniques:\code{\link[=resample_bootstrap(), resample_partition]{resample_bootstrap(), resample_partition()}}
}
