% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resample_bootstrap.R
\name{Resample_bootstrap}
\alias{Resample_bootstrap}
\title{bootstrap a resample object}
\usage{
Resample_bootstrap(data)
}
\arguments{
\item{data}{The \code{data} should be a data.frame}
}
\value{
resample
}
\description{
\code{Resample_bootstrap()} bootstrap the subset for a data.frame, using resample modal
class. Rows are sampled for data.
}
\examples{
b <- Resample_bootstrap(mtcars)
}
\seealso{
Other resampling techniques:\code{\link[=Resample(), Resample_partition]{Resample(), Resample_partition()}}
}
